+++ secret/cloud/selfconfig_aws.go
@@ -33,8 +33,8 @@ func init() {
 		if err != nil {
 			return nil, err
 		}
-		return selfConfigStoreAdapter{get: func(ctx context.Context, key string) (any, error) {
-			return store.GetSecret(ctx, key)
+		return selfConfigStoreAdapter{get: func(ctx context.Context, key string, secretOpts ...confii.SecretOption) (any, error) {
+			return store.GetSecret(ctx, key, secretOpts...)
 		}}, nil
 	})
 }